# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/customers": {
      "post": {
        "operationId": "CustomerCreateController_createUser",
        "summary": "",
        "description": "Create a new customer with permission or register customer",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      },
      "get": {
        "operationId": "CustomerController_getAll",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "id",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "phone",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "verified",
            "in": "query",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "required": false,
            "name": "active",
            "in": "query",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "get-all-customer": {
                    "summary": "Get All Customer",
                    "value": {
                      "data": [
                        {
                          "id": 1,
                          "name": "John Doe",
                          "email": "user@example.com",
                          "phone": "+123456789",
                          "verified": true,
                          "roleKey": "string",
                          "active": true,
                          "image": "https://example.com/avatar.png",
                          "createdAt": "2025-06-26T18:17:10.000Z",
                          "deletedAt": "2025-06-26T18:17:10.000Z",
                          "totalOrders": 1,
                          "totalSpent": 1
                        }
                      ],
                      "total": 1
                    }
                  },
                  "get-customer-with-id": {
                    "summary": "Get Customer with id",
                    "value": {
                      "data": {
                        "id": 1,
                        "name": "John Doe",
                        "email": "user@example.com",
                        "phone": "+123456789",
                        "verified": true,
                        "roleKey": "string",
                        "active": true,
                        "image": "https://example.com/avatar.png",
                        "createdAt": "2025-06-26T18:17:10.000Z",
                        "deletedAt": "2025-06-26T18:17:10.000Z",
                        "totalOrders": 1,
                        "totalSpent": 1
                      }
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      }
    },
    "/api/v1/customers/{id}": {
      "get": {
        "operationId": "CustomerController_getAll",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "id",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "phone",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "verified",
            "in": "query",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "required": false,
            "name": "active",
            "in": "query",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "get-all-customer": {
                    "summary": "Get All Customer",
                    "value": {
                      "data": [
                        {
                          "id": 1,
                          "name": "John Doe",
                          "email": "user@example.com",
                          "phone": "+123456789",
                          "verified": true,
                          "roleKey": "string",
                          "active": true,
                          "image": "https://example.com/avatar.png",
                          "createdAt": "2025-06-26T18:17:10.000Z",
                          "deletedAt": "2025-06-26T18:17:10.000Z",
                          "totalOrders": 1,
                          "totalSpent": 1
                        }
                      ],
                      "total": 1
                    }
                  },
                  "get-customer-with-id": {
                    "summary": "Get Customer with id",
                    "value": {
                      "data": {
                        "id": 1,
                        "name": "John Doe",
                        "email": "user@example.com",
                        "phone": "+123456789",
                        "verified": true,
                        "roleKey": "string",
                        "active": true,
                        "image": "https://example.com/avatar.png",
                        "createdAt": "2025-06-26T18:17:10.000Z",
                        "deletedAt": "2025-06-26T18:17:10.000Z",
                        "totalOrders": 1,
                        "totalSpent": 1
                      }
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      },
      "patch": {
        "operationId": "CustomerController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      },
      "delete": {
        "operationId": "CustomerController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Aroovia",
    "description": "",
    "version": "1.0",
    "contact": {
      "name": "Fahd Hakem",
      "url": "www.linkedin.com/in/fahd-hakem",
      "email": "fhakem75@gmail.com"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "ACCESS Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "REFRESH Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "PASSWORD_RESET Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "VERIFY Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "VISITOR Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCustomerDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "example": "test@test.com"
          },
          "phone": {
            "type": "string",
            "example": "+966 0509999999"
          },
          "password": {
            "type": "string",
            "example": "Default@123",
            "description": "Password must be: 8+ characters, 1 uppercase, 1 lowercase, 1 number, 1 special character"
          },
          "male": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "name",
          "email",
          "phone",
          "password",
          "male"
        ]
      },
      "UpdateCustomerDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "example": "test@test.com"
          },
          "phone": {
            "type": "string",
            "example": "+966 0509999999"
          },
          "male": {
            "type": "boolean",
            "example": true
          },
          "active": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  }
}